Force log.level to DEBUG if it is set to an incorrect value
Log::Log4perl will die on module load if log.level contains an incorrect value.
This is confusing to the user, especially in the Windows version, where nothing
is written to the console even if the app is started from it, and nothing is
written to the log file either.

If the value is incorrect, it will be forced to DEBUG, and a warning will be
logged to the log file.