Feature #2039: Properties improvements and fixes
Fixed a bug when getting the properties details of an object. A view is generated for each object in order to obtain the property list with its respective value, but due to the new table QVD_Object_Property_List, the link between the object property list and the property list were incorrect.