Feature #2140: Include support for SSL
The API and WAT were unified to be supplied in the port 80.
When the service is launched with hypnotoad no parameters are need. If morbo is used, the following argument must be provided: daemon -l 'https://*:80?cert=/etc/qvd/certs/server-certificate.pem&key=/etc/qvd/certs/server-private-key.pem
Added small fixes to the WAT, in order to work with the new API address.