Task #2749: Reset views does not work
Due to a change in the parse function of the JSON in Task #2453 the arguments were not being interpreted correctly. However, the arguments shall be read from the request instead of directly from the JSON. The function was simplified to use the request to delete the rows from the view tables.