Fixes #2854: Usage of smart-match operator was breaking QVD::DB::Translator
Besides that, the internal structure of this module was horrendous, with a
lot repeated code for argument checking.

Now, argument checking is performed by a couple of helper functions and
more strict checks are being performed on them.

Errors are reported as exceptions, instead of silently returning undef.

Usage of smart-matching has been eliminated.