Fixes #2874: Enable memory limits for LXC containers
Note that in order to limit the memory used by containers, the
kernel must be compiled with CONFIG_MEMCG_SWAP set and the feature
enabled at boot time passing the argument 'swapaccount=1' to the
kernel.

Anyway, the LXC backend detects when the feature is not available
and in that case doesn't try to set any limits.

This patch also adds support for autodetecting the cgroup control
directories when not configured.