Fixes Bug  #2855 and Bug #2863
Pool of events from Postgres database was managed with AnyEvent and generated a conflict with Mojolicious, it was replace with Mojo::Pg.

In addition, DB::Simple was used to connect to the database, because it is a singleton and this avoids the generation of several connections.

Useless package QVD::API::DBConfigProvider was removed.