Fixes #2876: Set per container cpusets
Now the HKD assigns a cpuset to every container at startup.

The configuration key `vm.lxc.cpuset.size` can be used to change
the size of the cpuset.

In the future that value should be sticked to the OSF or VM objects.

Besides that, this patch adds a new HKD agent, the hypervisor that
stores global information about the hypervisor as in this case, how
many containers are assigned to every CPU.