Task #3067: Migration procedure from 3.6 to 4.0
Deploy script was improved:
- Now it may accept more than one data file.
- Certificates data of the database was split in a different data file.
- Added a new flag --update-schema to the deploy script to update the current tuples of the database with the latest schema version.

The upgrade option generates a data file with the state of the current state of the database. Afterwards this file is parsed in a memory data structure. Finally the data is modified to be congruent with the actual database schema.

In addition some minor changes to clean up the code, as function name changes.