Bug #3358: ACLs API actions does not work
The actions "acl_get_list" and "get_acls_in_admins" were not receiving the correct argument type.

The actions "get_acls_in_admins" and "get_acls_in_roles" only require one argument and it was not being checked properly, so it was changed to ensure there is only one admin/role to get the acl list.