Bug #3085: Error while updating several elements with properties at the same time
Removed dependency of massive update function. These functions have some errors and shall be removed in the future because they do not provide any additional functionality. In addition, a bug was fixed that did that the update of one OSF was made several times, once per tag of associated DIs.