Fix QVD::Log default log path initialization
When QVD::Log is loaded and there is no set configuration about where
to put the log files, a number of default locations are tried. If the
name of the daemon is known, it is used, but otherwise it defaults to
using $0.

Apply basename() to $0 so that if $0 contains a path, it is removed.